feat: restart delay period
The restart delay period should allow the attempt to be reset if
the worker is restarted after that time. This fixes the issue where
we might see the attempt climbing even though the restarts over a
long period of time.

This is backwards compatible with the v4 version, as the restart
delay is a constant, so the attempt is disguarded.